+++ Changelog.md
@@ -17,6 +17,9 @@
 ### Misc Changes
 
 - [#675]: Minimum supported version of serde raised to 1.0.139
+- [#675]: Rework the `quick_xml::Error` type to provide more accurate information:
+  - `Error::UnexpectedBang` replaced by `SyntaxError`
+  - `Error::UnexpectedEof` replaced by `SyntaxError` in some cases
 
 [#675]: https://github.com/tafia/quick-xml/pull/675